Predicate removing met_when on it existing

+++ app/controllers/needs_controller.rb
@@ -239,7 +239,7 @@ class NeedsController < ApplicationController
       justifications: [],
       met_when: [],
     ).tap do |cleaned_params|
-      cleaned_params[:met_when].delete_if(&:empty?)
+      cleaned_params[:met_when].delete_if(&:empty?) if cleaned_params.key? :met_when
 
       %w(justifications organisation_ids).each do |field|
         cleaned_params[field].select!(&:present?) if cleaned_params[field]
